converted to typescript
