+++ tsconfig.json
@@ -2,7 +2,8 @@
   "compilerOptions": {
     "outDir": "./dist",
     "allowJs": true,
-    "target": "es5"
+    "target": "es5",
+    "downlevelIteration": true
   },
   "include": ["./webAO/*"]
 }
